% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abpackage.R
\docType{package}
\name{abpackage}
\alias{abpackage}
\alias{abpackage-package}
\title{abpackage: Pre-Period Enhanced A/B Testing}
\description{
Estimation and testing of percent change and
difference between treatment and control in the post-period
leveraging pre-period data.
}

