% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{CenterAndScalePrePeriod}
\alias{CenterAndScalePrePeriod}
\title{Center and scale the pre-period data according to weights.}
\usage{
CenterAndScalePrePeriod(data, weights)
}
\arguments{
\item{data}{A data.frame with 2 columns for each sample. The column condition
indicating whether the sample is from the control or treatment group.
A column pre indicating the observed value in the pre-period.}

\item{weights}{An data.frame with two variables and a single row.
The variables \code{control} and \code{treatment} indicate the weight of
the two groups.}
}
\description{
Center and scale the pre-period data according to weights.
}
\keyword{internal}

