% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{ComputeCI}
\alias{ComputeCI}
\title{Compute credible interval from a vector of data}
\usage{
ComputeCI(data, ci.level)
}
\arguments{
\item{data}{A vector.}

\item{ci.level}{The confidence level for the credible intervals.}
}
\value{
A data.frame with lower, center and upper values for the credible interval.
}
\description{
Compute credible interval from a vector of data
}

