% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{ComputePValue}
\alias{ComputePValue}
\title{Computes empirical p-value from vector of data}
\usage{
ComputePValue(data)
}
\arguments{
\item{data}{A vector.}
}
\value{
The p-value.
}
\description{
Computes empirical p-value from vector of data
}
\keyword{internal}

