% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{ComputeStats}
\alias{ComputeStats}
\title{Compute stats from a vector of data}
\usage{
ComputeStats(data, ci.level, which)
}
\arguments{
\item{data}{A vector.}

\item{ci.level}{The confidence level for the credible intervals.}

\item{which}{A string indicating the type of confidence interval.
Options are 'percent.change' and 'difference'.}
}
\value{
A data.frame with lower, median and upper values for the credible interval,
  mean, variance, p.value and type.
}
\description{
Compute stats from a vector of data
}
\keyword{internal}

