% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{ConditionalRegression}
\alias{ConditionalRegression}
\title{Computes a deterministic approximation of the second stage of pre-post.}
\usage{
ConditionalRegression(post.data, pre.data, mu.0, unif.nodes)
}
\arguments{
\item{post.data}{Vector of data in the experiment period.}

\item{pre.data}{vector of data in the pre-period.}

\item{mu.0}{Point estimate of the mean during the pre-period.}

\item{unif.nodes}{Vector of equally spaced nodes between zero and one.}
}
\value{
A discrete approximation of the posterior distribution of the mean.
}
\description{
Computes a deterministic approximation of the second stage of pre-post.
}

