% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{EmptyOutput}
\alias{EmptyOutput}
\title{Returns output where all values are NAs.}
\usage{
EmptyOutput()
}
\value{
A data.frame with 2 rows and 7 columns:
  lower, center, upper: the lower limit, the median and the upper
    limit of the credible interval.
  mean: the posterior mean.
  var: the posterior variance.
  p.value: the p-value.
  type: a string equal to percent.change or difference.
}
\description{
Returns output where all values are NAs.
}
\keyword{internal}

