% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.ab.R
\name{format.ab}
\alias{format.ab}
\title{Format an \code{ab} object for pretty printing.}
\usage{
format.ab(object, digits = 3, only.sig = FALSE, percent.change = TRUE,
  ...)
}
\arguments{
\item{object}{An object of class \code{ab}, usually, a result of a call to
\code{\link{PrePost}}.}

\item{digits}{Number of significant digits for the output.}

\item{only.sig}{If TRUE, returns only credible intervals associated to tests
that are statistically significant.}

\item{percent.change}{If TRUE, returns the credible intervals for the
percent change. If FALSE, returns the credible intervals for
the difference.}
}
\description{
Format an \code{ab} object for pretty printing.
}

