% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cis.R
\name{GetCIs}
\alias{GetCIs}
\title{Returns credible intervals from an object of class \code{ab}.}
\usage{
GetCIs(object, only.sig = FALSE, percent.change = TRUE, metrics)
}
\arguments{
\item{object}{An object of class \code{ab}.}

\item{only.sig}{If TRUE, returns only confidence intervals associated to
tests that are statistically significant after multiple testing correction.}

\item{percent.change}{If TRUE, returns credible intervals for the
percent change. If FALSE, returns credible intervals for
the difference.}

\item{metrics}{The names of the metrics to get. If missing, all metrics
are returned.}
}
\value{
A data.frame with the credible intervals,
  the point estimate and the metric name.
}
\description{
Returns credible intervals from an object of class \code{ab}.
}
\examples{
data <- SampleData(n.metrics = 20)
ans <- PrePost(data)
GetCIs(ans)
}

