% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{IsIllConditioned}
\alias{IsIllConditioned}
\title{Checks whether the design matrix for the second stage of PrePost is ill
conditioned.}
\usage{
IsIllConditioned(x, threshold = 1e-10)
}
\arguments{
\item{x}{Vector of pre-period data for either control or treatment group.}

\item{threshold}{The condition number threshold to determine if the design
matrix is ill conditioned.}
}
\description{
Checks whether the design matrix for the second stage of PrePost is ill
conditioned.
}

