% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{IsIllConditionedMatrix}
\alias{IsIllConditionedMatrix}
\title{Checks whether the design matrix for the second stage of PrePost is ill
conditioned.}
\usage{
IsIllConditionedMatrix(x.centered, threshold)
}
\arguments{
\item{x.centered}{Centered vector of pre-period data for either control or
treatment group.}

\item{threshold}{The condition number threshold to determine if the design
matrix is ill conditioned.}
}
\description{
Checks whether the design matrix for the second stage of PrePost is ill
conditioned.
}
\keyword{internal}

