% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{IsPrePeriodValid}
\alias{IsPrePeriodValid}
\title{Check whether to use the pre.period or not.}
\usage{
IsPrePeriodValid(data, weights, metric)
}
\arguments{
\item{data}{A data frame with columns pre and condition.}

\item{metric}{The name of the metric.}
}
\value{
A boolean indicating whether to use the pre-period or not.
}
\description{
Check whether to use the pre.period or not.
}
\keyword{internal}

