% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{NonStdStudentTQuantiles}
\alias{NonStdStudentTQuantiles}
\title{Computes the quantiles of the non-standardized Student's t-distribution.}
\usage{
NonStdStudentTQuantiles(unif.nodes, df, sigma, mu)
}
\arguments{
\item{unif.nodes}{Vector of probabilities.}

\item{df}{Degrees of freedom.}

\item{sigma}{Scale parameter.}

\item{mu}{Location paramter.}
}
\value{
A vector of quantiles.
}
\description{
Computes the quantiles of the non-standardized Student's t-distribution.
}
\keyword{internal}

