% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_check.R
\name{PermutationTest}
\alias{PermutationTest}
\title{Compute permutation p-value for mean difference between treatment and control}
\usage{
PermutationTest(data, n.samples = 50000)
}
\arguments{
\item{data}{A data.frame with 2 columns for each sample:
A column named "condition" indicating whether the sample is from
the control or treatment group.
A column named "pre" indicating the observed value in the pre-period.}

\item{n.samples}{Number of samples used for the permutation.}
}
\value{
A p-value associated to the null hypothesis that the two means are
  identical.
}
\description{
Compute permutation p-value for mean difference between treatment and control
}

