% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ab.R
\name{plot.ab}
\alias{plot.ab}
\title{Plot credible intervals}
\usage{
plot.ab(object, only.sig = FALSE, percent.change = TRUE, metrics,
  legend = FALSE, size = 1, horizontal = TRUE, n.char = 25, ...)
}
\arguments{
\item{object}{An object of class \code{ab}.}

\item{only.sig}{If TRUE, plots only credible intervals associated to tests
that are statistically significant.}

\item{percent.change}{If TRUE, plots the credible intervals for the
percent change. If FALSE, plots the credible intervals for the difference.}

\item{metrics}{The names of the metrics to plot. If missing, all metrics
are plotted.}

\item{legend}{If TRUE, plots a legend.}

\item{size}{Width of the credible intervals.}

\item{horizontal}{If TRUE, the credible intervals are horizontal.}
}
\value{
Plot of confidence intervals.
}
\description{
Plot credible intervals
}
\examples{
  data <- SampleData(n.metric = 3, mu.trmt = c(12, 10, 10))
  ans <- PrePost(data)
  plot(ans)
  plot(ans, horizontal = FALSE)
  plot(ans, only.sig = TRUE)
}

