% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{Posterior}
\alias{Posterior}
\title{Computes the posterior distribution of the post-period means.}
\usage{
Posterior(trmt.data, ctrl.data, weights, n.nodes, pre.period)
}
\arguments{
\item{trmt.data}{Data.frame for the treatment group.}

\item{ctrl.data}{Data.frame for the control group.}

\item{weights}{Data.frame with two variables and a single row.
The variables \code{control} and \code{treatment} indicate the weight of
the two groups.}

\item{n.nodes}{The number of nodes used to approximate each univariate
distribution.}

\item{pre.period}{Boolean indicating whether to use the pre-period or not.}
}
\value{
A matrix with two columns, representing the posterior of the control
   and the treatment group.
}
\description{
Computes the posterior distribution of the post-period means.
}
\keyword{internal}

