% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{PostOnlyPosterior}
\alias{PostOnlyPosterior}
\title{Computes the posterior distribution of the post-period means, when using
only post-period data.}
\usage{
PostOnlyPosterior(trmt.data, ctrl.data, unif.nodes)
}
\arguments{
\item{trmt.data}{Vector of post-period data from the treatment group.}

\item{ctrl.data}{Vector of post-period data from the control group.}

\item{unif.nodes}{Vector of probabilities.}
}
\value{
A matrix with two columns, representing the posterior of the control
   and the treatment group.
}
\description{
Computes the posterior distribution of the post-period means, when using
only post-period data.
}
\keyword{internal}

