% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_check.R
\name{PreCheck}
\alias{PreCheck}
\title{Check pre-period alignment between treatment and control}
\usage{
PreCheck(data, n.samples = 50000)
}
\arguments{
\item{data}{A data.frame with several variables. The variable
\code{condition} indicates the treatment assignment (\code{control} or
\code{treatment}) for each row. The variable \code{pre} indicates
the observed value during the pre-period for each row.
The variable \code{metric} indicates the name of each metric.
If omitted, all rows are assumed to be from the same metric.}

\item{n.samples}{Number of samples used for the permutation.}
}
\value{
A data.frame with a column metric, a column p-value and a column
  misalignemnt indicating the level of misalignemnt between the treatment
  and the control.
}
\description{
Check pre-period alignment between treatment and control
}
\details{
If all metrics are independent, 5% of metrics are expected
   to be classified as * (light misalignment), 4% of metrics are expected to
   be classified as ** (medium misalignment), and 1% of metrics are
   expected to be classified as *** (heavy misalignment). P-values are
   computed using \code{\link[coin]{oneway_test}}.
}
\examples{
  data <- SampleData(n.metrics = 20)
  (ans <- PreCheck(data))
}

