% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_post.R
\name{PrePost}
\alias{PrePost}
\title{Compute credible intervals}
\usage{
PrePost(data, weights = NULL, ci.level = 0.95, p.threshold = 0.05,
  p.method = "none", n.nodes = 50)
}
\arguments{
\item{data}{A data.frame with several variables. The variable
\code{condition} indicates the treatment assignment (\code{control} or
\code{treatment}) for each row. The variable \code{post} indicates
the observed value during the experiment period for each row. The variable
\code{pre} indicates the observed value during the pre-period for each row.
If omitted, the pre-period is not used to estimate the treatment effect.
The variable \code{metric} indicates the name of each metric.
If omitted, all rows are assumed to be from the same metric.}

\item{weights}{An optional data.frame with two variables and a single row.
The variables \code{control} and \code{treatment} indicate the weight of
the two groups. The weight must be proportional to the traffic proportion
of observations in the condition group.
If NULL, equal weights are are used in the fitting process.}

\item{ci.level}{Confidence level for credible intervals.}

\item{p.threshold}{P-value threshold used to identify statistically
significant effects. See \code{\link[stats]{p.adjust}} function for more
details.}

\item{p.method}{Correction method for multiple comparison. See
\code{\link[stats]{p.adjust}} function for more details.}

\item{n.nodes}{Number of nodes used for estimation.}
}
\value{
An object of class \code{ab}.
}
\description{
Compute credible intervals
}
\details{
Compute percent change and difference estimation between treatment
  and control for one or several metrics.
}
\examples{
data <- SampleData(n.metrics = 20)
(ans <- PrePost(data))
plot(ans)

data.no.pre <- dplyr::select(data, -pre)
(ans.no.pre <- PrePost(data.no.pre))
plot(ans.no.pre)
}

