% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{PrePostPosterior}
\alias{PrePostPosterior}
\title{Computes the posterior distribution of the post-period means, when
correcting for the pre-period.}
\usage{
PrePostPosterior(trmt.data, ctrl.data, weights, unif.nodes)
}
\arguments{
\item{trmt.data}{Data.frame for the treatment group.}

\item{ctrl.data}{Data.frame for the control group.}

\item{weights}{Data.frame with two variables and a single row.
The variables \code{control} and \code{treatment} indicate the weight of
the two groups.}

\item{unif.nodes}{Vector of probabilities.}
}
\value{
A matrix with two columns, representing the posterior of the control
   and the treatment group.
}
\description{
Computes the posterior distribution of the post-period means, when
correcting for the pre-period.
}
\keyword{internal}

