% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{ReshapeData}
\alias{ReshapeData}
\title{Reshape data to have the structure required by \code{\link{PrePost}} and
\code{\link{PreCheck}}}
\usage{
ReshapeData(data, condition.col = "condition",
  observation.col = "observation", pre.post.col = "pre.post",
  condition.levels = c("control", "treatment"), pre.post.levels = c("pre",
  "post"))
}
\arguments{
\item{data}{A data.frame with several variables. The variable
\code{condition} indicates the treatment assignment (\code{control} or
\code{treatment}) for each row. The variable \code{observation} indicates
the observation number of each row. The variable \code{pre.post} indicates
the period (\code{pre} or \code{post}) for each row.
The remaining variables represent the metrics to be tested. If there
are multiple rows for a given combination of \code{condition},
\code{observation} and \code{pre.post}, these rows are summed up.}

\item{condition.col}{The name of the variable indicating the treatment
assignment for each row in \code{data}.}

\item{observation.col}{The name of the variable indicating the observation
for each row in \code{data}.}

\item{pre.post.col}{The name of the variable indicating the period for each
row in \code{data}.}

\item{condition.levels}{The names indicating \code{control} and
\code{treatment} in \code{data}.}

\item{pre.post.levels}{The names indicating \code{pre} and \code{post} period
in \code{data}.}
}
\value{
A data.frame with several variables. The variable \code{condition}
  indicates the treatment assignment (\code{control} or \code{treatment})
  for each row. The variable \code{pre} indicates
  the observed value during the pre-period for each row.
  The variable \code{metric} indicates the name of each metric.
}
\description{
Reshape data to have the structure required by \code{\link{PrePost}} and
\code{\link{PreCheck}}
}
\examples{
  data <- SampleData(n.metrics = 4, spread = TRUE)
  reshaped.data <- ReshapeData(data)
  ans <- PrePost(reshaped.data)
}

