% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{SampleData}
\alias{SampleData}
\title{Generates synthetic data from Gaussian distributions}
\usage{
SampleData(n.metrics = 100, n.observations = 20, mu.pre = 100,
  sigma.pre = 1, rho.ctrl = 0.8, rho.trmt = NULL, mu.ctrl = NULL,
  sigma.ctrl = NULL, mu.trmt = NULL, sigma.trmt = NULL, spread = FALSE)
}
\arguments{
\item{n.metrics}{Number of metrics.}

\item{n.observations}{Number of observations.}

\item{mu.pre}{Vector or scalar representing the mean during the pre-period
for each metric.}

\item{sigma.pre}{Vector or scalar representing the standard deviation during
the pre-period for each metric.}

\item{rho.ctrl}{Vector or scalar representing the correlation between the
pre-period and the post-period for the control group for each metric.}

\item{rho.trmt}{Vector or scalar representing the correlation between the
pre-period and the post-period for the treatment group for each metric.
If missing, set equal to rho.ctrl.}

\item{mu.ctrl}{Vector or scalar representing the mean during the post-period
for the control group for each metric. If missing, set equal to mu.pre.}

\item{sigma.ctrl}{Vector or scalar representing the standard deviation during
the post-period for the control group for each metric. If missing, set
equal to sigma.pre.}

\item{mu.trmt}{Vector or scalar representing the mean during the post-period
for the treatment group for each metric. If missing, set equal to mu.pre.}

\item{sigma.trmt}{Vector or scalar representing the standard deviation
during the post-period for the treatment group for each metric. If missing,
set equal to sigma.pre.}

\item{spread}{If TRUE, the output has a horizontal structure
with a \code{pre.post} column and column for each metric.
If FALSE, the output has a vertical structure with a
\code{metric} column, a \code{pre} column and a \code{post} column.}
}
\value{
A \code{data.frame} which can be used as input for the \code{PrePost}
  and the \code{PreCheck} functions.
}
\description{
Generates synthetic data from Gaussian distributions
}
\examples{
  data <- SampleData(n.metrics = 10)
  ans <- PrePost(data)

  n.metrics <- 10
  mu.pre <- 100
  mu.ctrl <- 100
  mu.trmt <- c(110, rep(100, n.metrics - 1))
  data <- SampleData(n.metrics = n.metrics,
                     mu.pre = mu.pre,
                     mu.ctrl = mu.ctrl,
                     mu.trmt = mu.trmt)
  ans <- PrePost(data)

  data <- SampleData(n.metrics = n.metrics,
                     mu.pre = mu.pre,
                     mu.ctrl = mu.ctrl,
                     mu.trmt = mu.trmt,
                     spread = TRUE)
  std.data <- ReshapeData(data)
  ans <- PrePost(std.data)
}

