% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{SecondStageEstimation}
\alias{SecondStageEstimation}
\title{Computes estimates of the experiment period means given the pre-period mean.}
\usage{
SecondStageEstimation(mu.0, trmt.post.data, ctrl.post.data, trmt.pre.data,
  ctrl.pre.data, unif.nodes)
}
\arguments{
\item{mu.0}{The mean in the pre-period.}

\item{trmt.post.data}{A vector of data under treatment in the experiment
period.}

\item{ctrl.post.data}{A vector of data under control in the experiment
period.}

\item{trmt.pre.data}{A vector of data under treatment in the pre-period.}

\item{ctrl.pre.data}{A vector of data under control in the pre-period.}

\item{unif.nodes}{A vector of equally spaced nodes between zero and one.}
}
\value{
A matrix with posterior estimates of the experiment period means.
}
\description{
Computes estimates of the experiment period means given the pre-period mean.
}

