% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{UnivariateLinearRegression}
\alias{UnivariateLinearRegression}
\title{Computes the parameters of a univariate linear regression.}
\usage{
UnivariateLinearRegression(y, x)
}
\arguments{
\item{y}{Dependent variable.}

\item{x}{Explanatory variable.}
}
\value{
A data.frame with:
    intercept: The estimate of the intercept coefficient.
    df: The degress of freedom of the standard deviation of the residuals.
    sigma: The standard deviation of the residuals.
    XtX.inverse.intercept: The top left elemeng of the inverse
      of the matrix X'X.
}
\description{
Computes the parameters of a univariate linear regression.
}
\keyword{internal}

