% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pre_post.R
\name{WeightedAverage}
\alias{WeightedAverage}
\title{WeightedAverage returns the weighted average of two vectors}
\usage{
WeightedAverage(x.1, x.2, weight.1 = 1, weight.2 = 1)
}
\arguments{
\item{x.1}{A numerical vector.}

\item{x.2}{A numerical vector.}

\item{weight.1}{A positive weight.}

\item{weight.2}{A positive weight.}
}
\value{
The weighted average of x.1 and x.2.
}
\description{
WeightedAverage returns the weighted average of two vectors
}
\keyword{internal}

